+++ src/main/java/com/wanted/preonboardingbackend/recruit/dto/RecruitCreateRequest.java
@@ -10,7 +10,7 @@ public class RecruitCreateRequest {
     
     String jobPosition;
 
-    int signingBonus;
+    long signingBonus;
 
     String jobDescription;
